+++ src/taiko/skills/traits.rs
@@ -48,9 +48,11 @@ pub(crate) trait StrainSkill: Skill {
             if self.curr_section_peak().abs() <= f64::EPSILON
                 && curr.start_time > *self.curr_section_end()
             {
-                let remaining = curr.start_time - *self.curr_section_end();
-                let skip_iter_count = (remaining / SECTION_LEN).ceil();
-                *self.curr_section_end() += skip_iter_count * SECTION_LEN;
+                let remaining_time = curr.start_time - *self.curr_section_end();
+                let remaining_iters = (remaining_time / SECTION_LEN).ceil();
+                *self.curr_section_end() += remaining_iters * SECTION_LEN;
+                self.strain_peaks_mut()
+                    .push_n_zeros(remaining_iters as usize);
             }
         }